pyproject.toml
[tool.poetry]
name = "oralb-ble"
version = "0.17.3"
description = "Bluetooth library for Oral B devices"
authors = ["J. Nick Koston <bluetooth@koston.org>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Bluetooth-Devices/oralb-ble"
documentation = "https://oralb-ble.readthedocs.io"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]
packages = [
    { include = "oralb_ble", from = "src" },
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Bluetooth-Devices/oralb-ble/issues"
"Changelog" = "https://github.com/Bluetooth-Devices/oralb-ble/blob/main/CHANGELOG.md"

[tool.poetry.dependencies]
python = "^3.9"

# Documentation Dependencies
Sphinx = {version = "^5.0", optional = true}
sphinx-rtd-theme = {version = "^1.0", optional = true}
myst-parser = {version = "^0.18", optional = true}
bluetooth-sensor-state-data = ">=1.6.0"
bluetooth-data-tools = ">=0.1.2"
bleak = ">=0.19.2"
home-assistant-bluetooth = ">=1.6.0"
bleak-retry-connector = ">=2.13.0"

[tool.poetry.extras]
docs = [
    "myst-parser",
    "sphinx",
    "sphinx-rtd-theme",
]

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^3.0"

[tool.poetry.group.dev.dependencies]
mock = ">=5.0.0"
pytest-asyncio = ">=0.20.0"

[tool.semantic_release]
branch = "main"
version_toml = "pyproject.toml:tool.poetry.version"
version_variable = "src/oralb_ble/__init__.py:__version__"
build_command = "pip install poetry && poetry build"

[tool.pytest.ini_options]
addopts = "-v -Wdefault --cov=oralb_ble --cov-report=term-missing:skip-covered"
pythonpath = ["src"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@overload",
    "if TYPE_CHECKING",
    "raise NotImplementedError",
]

[tool.isort]
profile = "black"
known_first_party = ["oralb_ble", "tests"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src/"
no_implicit_optional = true
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
exclude = [
    'docs/.*',
    'setup.py',
]

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "docs.*"
ignore_errors = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
